refactor: route group layout 적용
