+++ app/(search)/layout.tsx
@@ -1,0 +1,12 @@
+export default function SearchLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <>
+      <div>search header</div>
+      {children}
+    </>
+  );
+}
